skillforge-backend/test-frontend/pyproject.toml
[project]
name = "test-frontend"
version = "0.1.0"
description = "SkillForge Streamlit Frontend - AI-powered learning assistant"
authors = [{ name = "SkillForge Team", email = "team@skillforge.dev" }]
readme = "README.md"
keywords = ['python', 'streamlit', 'ai', 'learning', 'education']
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://fpgmaas.github.io/test-frontend/"
Repository = "https://github.com/fpgmaas/test-frontend"
Documentation = "https://fpgmaas.github.io/test-frontend/"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.9.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # ComplexityTooHigh - Allow complex functions
    "C901",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
